requirements.txt
streamlit==1.39.0
pandas==2.2.3
numpy==1.26.4
plotly==5.24.1
matplotlib==3.9.2
wordcloud==1.9.3
python-dotenv==1.0.1
requests==2.32.3
pyjwt==2.9.0
cryptography==43.0.1
nltk==3.9.1
langdetect==1.0.9
openai==1.50.2
google-api-python-client==2.149.0
transformers==4.44.2
sentence-transformers==3.1.1
torch==2.4.1
faiss-cpu==1.8.0.post1
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
scikit-learn==1.5.2
httpx==0.27.0